stage4/pyproject.toml
[project]
name = "stage4-inference"
version = "0.1.0"
description = "Stage 4: Inference and Auto-Annotation pipeline stage"
requires-python = ">=3.10"
dependencies = [
    "pipeline-common @ file://${PROJECT_ROOT}/pipeline_common",
    "boto3>=1.34.0",
    "botocore>=1.34.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "mlflow>=2.8.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "opencv-python-headless>=4.8.0",
    "Pillow>=10.0.0",
    "flask>=3.0.0",
    "prometheus-client>=0.19.0",
]
optional-dependencies = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.10"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
